+++ Editor/Rendering/GraphicsSettingsLoader.cs
@@ -1,5 +1,9 @@
 using UnityEditor;
 using UnityEngine;
+using UnityEditor.XR.CompositionLayers.Editor.Emulation;
+using Unity.XR.CompositionLayers.Services;
+using Unity.XR.CompositionLayers.Emulation;
+using Unity.XR.CompositionLayers.Extensions;
 
 namespace Unity.XR.CompositionLayers.Rendering.Editor
 {
@@ -15,7 +19,8 @@ namespace Unity.XR.CompositionLayers.Rendering.Editor
             if ((EditorUserBuildSettings.activeBuildTarget == BuildTarget.StandaloneWindows ||
                 EditorUserBuildSettings.activeBuildTarget == BuildTarget.StandaloneWindows64 ||
                 EditorUserBuildSettings.activeBuildTarget == BuildTarget.StandaloneOSX ||
-                EditorUserBuildSettings.activeBuildTarget == BuildTarget.StandaloneLinux64) && CompositionLayersRuntimeSettings.Instance.EmulationInStandalone)
+                EditorUserBuildSettings.activeBuildTarget == BuildTarget.StandaloneLinux64) &&
+                CompositionLayersRuntimeSettings.Instance.EmulationInStandalone)
             {
                 AddShadersToGraphicsSettings();
             }
@@ -27,6 +32,12 @@ namespace Unity.XR.CompositionLayers.Rendering.Editor
             {
                 Debug.Log("Standalone build target detected. Adding Uber shader to Graphics Settings for Emulation.");
             }
+
+            if (GraphicsSettingsHelpers.AddAlwaysIncludedShaders(GraphicsSettingsHelpers.ShaderType.ColorScaleBias))
+            {
+                Debug.Log("Standalone build target detected. Adding Color Scale Bias shader to Graphics Settings for Emulation.");
+            }
+
         }
     }
 }